# services/orchestrator-api/src/main/java/com/orquestador/orchestrator/api/dtos/ErrorResponse.java
package com.orquestador.orchestrator.api.dtos;

import lombok.AllArgsConstructor;
import lombok.Getter;

import java.time.LocalDateTime;
import java.util.List;

@Getter
@AllArgsConstructor
public class ErrorResponse {
    private String code;
    private String message;
    private LocalDateTime timestamp;
    private List<String> details;
}